update router structure
