+++ server/routes/userRouter.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const userController = require('../controllers/clientController');
-const jwt = require('jsonwebtoken');
+const userController = require('../controllers/userController');
 
 router.post('/signup', userController.signup, async (req, res) => {
   return res.status(200).json(res.locals.newClient);
@@ -14,13 +13,5 @@ router.post('/login', userController.login, async (req, res) => {
     return res.status(200).header("auth-token", res.locals.result.token).json(res.locals.result);
   }
 });
-
-router.post('/newInstance', userController.verifyToken, userController.createInstance, (req, res) => {
-  return res.status(200).json(res.locals.instance)
-});
-
-router.post('/test', userController.verifyToken, (req, res) => {
-  return res.status(200).json(res.locals.client)
-})
 
 module.exports = router;
